+++ Core/Timer.cpp
@@ -12,10 +12,10 @@
 
 namespace Time
 {
-    void time::Sleep(uint64_t Milliseconds)
+    void time::Sleep(uintptr_t Milliseconds)
     {
 #if defined(__amd64__) || defined(__i386__)
-        uint64_t Target = mminq(&((HPET *)hpet)->MainCounterValue) + (Milliseconds * 1000000000) / clk;
+        uintptr_t Target = mminq(&((HPET *)hpet)->MainCounterValue) + (Milliseconds * 1000000000) / clk;
         while (mminq(&((HPET *)hpet)->MainCounterValue) < Target)
             CPU::Pause();
 #elif defined(__aarch64__)
